+++ src/main/java/com/space/server/ai/service/implementation/PromptCreator.java
@@ -1,7 +1,7 @@
 package com.space.server.ai.service.implementation;
 
 import com.space.server.chat.domain.value.Type;
-import com.space.server.core.chapter.domain.Chapter;
+import com.space.server.domin.chapter.domain.Chapter;
 import com.space.server.core.checklist.domain.Checklist;
 import com.space.server.core.quiz.domain.Quiz;
 import lombok.RequiredArgsConstructor;
