+++ src/main/java/com/space/server/chat/service/CommandChatService.java
@@ -15,8 +15,8 @@ import com.space.server.chat.service.implementation.ChatCreator;
 import com.space.server.chat.service.implementation.ChatReader;
 import com.space.server.chat.service.implementation.AiResponseJsonParsing;
 import com.space.server.chat.service.implementation.ChatValidator;
-import com.space.server.core.chapter.domain.Chapter;
-import com.space.server.core.chapter.service.implementation.ChapterReader;
+import com.space.server.domin.chapter.domain.Chapter;
+import com.space.server.domin.chapter.service.implementation.ChapterReader;
 import com.space.server.core.checklist.domain.Checklist;
 import com.space.server.core.checklist.service.implementation.ChecklistReader;
 import com.space.server.core.quiz.domain.Quiz;
