+++ src/main/java/com/space/server/core/quiz/domain/Quiz.java
@@ -1,6 +1,6 @@
 package com.space.server.core.quiz.domain;
 
-import com.space.server.core.chapter.domain.Chapter;
+import com.space.server.domin.chapter.domain.Chapter;
 import com.space.server.core.quiz.domain.value.CharacterDirection;
 import io.hypersistence.utils.hibernate.type.array.IntArrayType;
 import io.hypersistence.utils.hibernate.type.basic.PostgreSQLHStoreType;
