+++ src/main/java/com/space/server/core/quiz/domain/repository/QuizRepository.java
@@ -1,6 +1,6 @@
 package com.space.server.core.quiz.domain.repository;
 
-import com.space.server.core.chapter.domain.Chapter;
+import com.space.server.domin.chapter.domain.Chapter;
 import com.space.server.core.quiz.domain.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
