+++ src/main/java/com/space/server/core/quiz/presentation/dto/response/QuizResponse.java
@@ -1,6 +1,6 @@
 package com.space.server.core.quiz.presentation.dto.response;
 
-import com.space.server.core.chapter.presentation.dto.response.ChapterResponse;
+import com.space.server.domin.chapter.presentation.dto.response.ChapterResponse;
 import com.space.server.core.quiz.domain.Quiz;
 import com.space.server.core.quiz.domain.value.CharacterDirection;
 import io.swagger.v3.oas.annotations.media.Schema;
