+++ src/main/java/com/space/server/core/quiz/service/implementation/QuizReader.java
@@ -1,6 +1,6 @@
 package com.space.server.core.quiz.service.implementation;
 
-import com.space.server.core.chapter.domain.Chapter;
+import com.space.server.domin.chapter.domain.Chapter;
 import com.space.server.core.quiz.domain.Quiz;
 import com.space.server.core.quiz.domain.repository.QuizRepository;
 import com.space.server.core.quiz.exception.QuizNotFoundException;
