+++ src/main/java/com/space/server/domin/chapter/domain/repository/ChapterRepository.java
@@ -1,6 +1,6 @@
-package com.space.server.core.chapter.domain.repository;
+package com.space.server.domin.chapter.domain.repository;
 
-import com.space.server.core.chapter.domain.Chapter;
+import com.space.server.domin.chapter.domain.Chapter;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
