+++ src/main/java/com/space/server/domin/chapter/service/CommandChapterService.java
@@ -1,10 +1,10 @@
-package com.space.server.core.chapter.service;
+package com.space.server.domin.chapter.service;
 
-import com.space.server.core.chapter.domain.Chapter;
-import com.space.server.core.chapter.service.implementation.ChapterCreator;
-import com.space.server.core.chapter.service.implementation.ChapterDeleter;
-import com.space.server.core.chapter.service.implementation.ChapterReader;
-import com.space.server.core.chapter.service.implementation.ChapterUpdater;
+import com.space.server.domin.chapter.domain.Chapter;
+import com.space.server.domin.chapter.service.implementation.ChapterCreator;
+import com.space.server.domin.chapter.service.implementation.ChapterDeleter;
+import com.space.server.domin.chapter.service.implementation.ChapterReader;
+import com.space.server.domin.chapter.service.implementation.ChapterUpdater;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
