+++ src/main/java/com/space/server/domin/chapter/service/implementation/ChapterDeleter.java
@@ -1,7 +1,7 @@
-package com.space.server.core.chapter.service.implementation;
+package com.space.server.domin.chapter.service.implementation;
 
-import com.space.server.core.chapter.domain.Chapter;
-import com.space.server.core.chapter.domain.repository.ChapterRepository;
+import com.space.server.domin.chapter.domain.Chapter;
+import com.space.server.domin.chapter.domain.repository.ChapterRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
