+++ src/main/java/com/space/server/domin/chapter/service/implementation/ChapterReader.java
@@ -1,8 +1,8 @@
-package com.space.server.core.chapter.service.implementation;
+package com.space.server.domin.chapter.service.implementation;
 
-import com.space.server.core.chapter.domain.Chapter;
-import com.space.server.core.chapter.domain.repository.ChapterRepository;
-import com.space.server.core.chapter.exception.ChapterNotFoundException;
+import com.space.server.domin.chapter.domain.Chapter;
+import com.space.server.domin.chapter.domain.repository.ChapterRepository;
+import com.space.server.domin.chapter.exception.ChapterNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
